+++ example/lib/screens/component/apple_pay/apple_pay_advanced_component_screen.dart
@@ -1,7 +1,5 @@
 // ignore_for_file: unused_local_variable
 
-import 'dart:convert';
-
 import 'package:adyen_checkout/adyen_checkout.dart';
 import 'package:adyen_checkout_example/config.dart';
 import 'package:adyen_checkout_example/repositories/adyen_apple_pay_component_repository.dart';
@@ -45,14 +43,14 @@ class ApplePayAdvancedComponentScreen extends StatelessWidget {
       applePayConfiguration: _createApplePayConfiguration(),
     );
 
-    return FutureBuilder<String>(
+    return FutureBuilder<Map<String, dynamic>>(
       future: repository.fetchPaymentMethods(),
-      builder: (BuildContext context, AsyncSnapshot<String> snapshot) {
+      builder:
+          (BuildContext context, AsyncSnapshot<Map<String, dynamic>> snapshot) {
         if (snapshot.hasData) {
-          final AdvancedCheckoutPreview advancedCheckout =
-              AdvancedCheckoutPreview(
-                  onSubmit: repository.onSubmit,
-                  onAdditionalDetails: repository.onAdditionalDetailsMock);
+          final AdvancedCheckout advancedCheckout = AdvancedCheckout(
+              onSubmit: repository.onSubmit,
+              onAdditionalDetails: repository.onAdditionalDetailsMock);
           final paymentMethod = _extractPaymentMethod(snapshot.data!);
 
           return Column(
@@ -151,13 +149,13 @@ class ApplePayAdvancedComponentScreen extends StatelessWidget {
     );
   }
 
-  Map<String, dynamic> _extractPaymentMethod(String paymentMethods) {
+  Map<String, dynamic> _extractPaymentMethod(
+      Map<String, dynamic> paymentMethods) {
     if (paymentMethods.isEmpty) {
       return <String, String>{};
     }
 
-    Map<String, dynamic> jsonPaymentMethods = jsonDecode(paymentMethods);
-    return jsonPaymentMethods["paymentMethods"].firstWhere(
+    return paymentMethods["paymentMethods"].firstWhere(
       (paymentMethod) => paymentMethod["type"] == "applepay",
       orElse: () => throw Exception("Apple pay payment method not provided"),
     );