+++ lib/src/drop_in/drop_in.dart
@@ -82,17 +82,21 @@ class DropIn {
 
   Future<PaymentResult> startDropInAdvancedFlowPayment(
     DropInConfiguration dropInConfiguration,
-    String paymentMethodsResponse,
+    Map<String, dynamic> paymentMethodsResponse,
     Checkout advancedCheckout,
   ) async {
     adyenLogger.print("Start Drop-in advanced flow");
     final dropInAdvancedFlowCompleter = Completer<PaymentResultDTO>();
     final sdkVersionNumber =
         await sdkVersionNumberProvider.getSdkVersionNumber();
+    final encodedPaymentMethodsResponse = jsonEncode(
+      paymentMethodsResponse,
+      toEncodable: (value) => throw Exception("Could not encode $value"),
+    );
 
     dropInPlatformApi.showDropInAdvanced(
       dropInConfiguration.toDTO(sdkVersionNumber),
-      paymentMethodsResponse,
+      encodedPaymentMethodsResponse,
     );
 
     dropInFlutterApi.dropInAdvancedFlowPlatformCommunicationStream =
@@ -231,11 +235,6 @@ class DropIn {
     final Map<String, dynamic> submitDataDecoded = jsonDecode(submitData);
     switch (advancedCheckout) {
       case AdvancedCheckout it:
-        final paymentData = submitDataDecoded[Constants.submitDataKey];
-        final paymentDataEncoded = jsonEncode(paymentData);
-        final PaymentEvent paymentEvent = await it.onSubmit(paymentDataEncoded);
-        return paymentEvent;
-      case AdvancedCheckoutPreview it:
         final PaymentEvent paymentEvent = await it.onSubmit(
           submitDataDecoded[Constants.submitDataKey],
           submitDataDecoded[Constants.submitExtraKey],
@@ -250,8 +249,6 @@ class DropIn {
       PlatformCommunicationModel event, Checkout advancedCheckout) async {
     switch (advancedCheckout) {
       case AdvancedCheckout it:
-        return it.onAdditionalDetails(event.data as String);
-      case AdvancedCheckoutPreview it:
         final additionalDetails = jsonDecode(event.data as String);
         return await it.onAdditionalDetails(additionalDetails);
       case SessionCheckout():