+++ lib/src/drop_in/model/drop_in_configuration.dart
@@ -4,7 +4,7 @@ import 'package:adyen_checkout/src/common/model/payment_method_configurations/ap
 import 'package:adyen_checkout/src/common/model/payment_method_configurations/card_configuration.dart';
 import 'package:adyen_checkout/src/common/model/payment_method_configurations/cash_app_pay_configuration.dart';
 import 'package:adyen_checkout/src/common/model/payment_method_configurations/google_pay/google_pay_configuration.dart';
-import 'package:adyen_checkout/src/common/model/stored_payment_method_configuration.dart';
+import 'package:adyen_checkout/src/common/model/payment_method_configurations/stored_payment_method_configuration.dart';
 
 final class DropInConfiguration extends BaseConfiguration {
   final CardConfiguration? cardConfiguration;