+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## 0.0.4 (in development)
 
 * ⚠ **Breaking Change**: Changed ResultCode into an enum. Please adjust your usage of the PaymentResult (PaymentSessionFinished or PaymentAdvancedFinished). 
+* ⚠ **Breaking Change**: Change the paymentMethodsResponse type from String to Map<String, dynamic>.  
 * Updated iOS SDK to v5.8.0.
 * Updated Android SDK to v5.3.1.
 * Added support for customizing Drop-in and card component.
 * Added option to adjust Drop-in preselected payment method title.
+* Replaced advanced checkout class with the preview implementation.
 
 ## 0.0.3